ensure_units moved to utils.units
