+++ src/earthkit/climate/indicators/precipitation.py
@@ -12,9 +12,10 @@ from earthkit.climate.utils.conversions import (
     EarthkitData,
     MetadataDict,
     to_earthkit_field,
-    to_xarray_dataset, ensure_units,
+    to_xarray_dataset,
 )
 from earthkit.climate.utils.provenance import add_indicator_provenance
+from earthkit.climate.utils.units import ensure_units
 
 
 def maximum_consecutive_wet_days(
